ConsoleWindowController testability enhancement.
Make ConsoleWindowController possible (easier) to instantiate in a unit test.
Demonstration of a simple test that benefits from this change.

The change in ConsoleWindowController is one of several ways to make it easier to create an instance, without unnecessary external requirements. Perhaps not the best way. The main problem is with the Stage class, wich is not easy to create an instance of. Therefore it's no longer done when ConsoleWindowController is created.