Update module-info.java
All exports are removed from the module-info.java file, because it's not likely that this app needs to export anything.
Two opens statements are added, to try to fix behavior which otherwise changes.
The programs seems to work in my limited test, but the behavior needs to be checked by someone else.