Update MainLoader.java
The thread that creates a WebBrowserController and a DropboxViewer is named and assigned to a variable. The start() method call is put on a separate line, which can easily be commented out.

If browserThread.start(); is commented out, the thread isn't started, and I don't get the exceptions associated with the JxBrowser.
On the other hand, the Vacuum + Exit fails; it seems to be cleaning something forever, until the application is forced to quit.

Therefore, I haven't commented out start(). But I would like to, because the JxBrowser dependent functionality is not ready yet. It's better to write some unit tests that demonstrate a working WebBrowserController and a DropboxViewer before this thread is re-enabled.